+++ versions.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     uptime = {
       source  = "uptime-com/uptime"
-      version = ">= 2.10"
+      version = ">= 2.28"
     }
   }
 }